+++ services/restapi/src/main/java/com/jobsity/challenge/restapi/RestApi.kt
@@ -17,19 +17,19 @@ sealed interface RestApi {
      * @param showId Show ID on TVMaze
      * @return [TVShowDetail] object containing all details about a show
      */
-    suspend fun getShow(showId: Int): TVShowDetail
+    suspend fun getShow(showId: Long): TVShowDetail
 
     /**
      * Return the list of episodes for a given series separated by season
      * @param showId Show ID on TVMaze
      * @return A list of episodes for each season
      */
-    suspend fun getEpisodes(showId: Int): List<Episode>
+    suspend fun getEpisodes(showId: Long): List<Episode>
 
     /**
      * Return detailed info about an episode
      * @param episodeId Episode ID on TVMaze
      * @return [EpisodeDetail] object containing all details about an episode
      */
-    suspend fun getEpisode(episodeId: Int): EpisodeDetail
+    suspend fun getEpisode(episodeId: Long): EpisodeDetail
 }
